refactor: update project structure for tests
