+++ src/env.ts
@@ -14,9 +14,6 @@ const envSchema = z.object({
   USER_TOKEN: z.string(),
   SUPPLIER_TOKEN: z.string(),
   URL:  z.string().url(),
-  OWNER_USER: z.string(),
-  OWNER_EMAIL: z.string().email(),
-  OWNER_PASSWORD: z.string()
 });
 
 const env = envSchema.parse(process.env);
